pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "template"
version = "0.0.1"
authors = [
  { name="Mehmet Yasin Cifci", email="myasincifci@gmail.com" },
]
description = "Template for DL projects"
readme = "README.md"
requires-python = "==3.11.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'torch>=2.0',
  'torchvision>=0.17.2',
  'pytorch-lightning>=2.2.2',
  'torchmetrics>=1.3.2',
  'hydra-core>=1.3.2',
  'wandb>=0.16.6',
]

[project.urls]
Homepage = "https://github.com/myasincifci/template"
Issues = "https://github.com/myasincifci/template/issues"